+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Hotfixes (`0.1.2+1.21`)
+## Added
+ - Zombified Piglins have a 10% chance of spawning left-handed.
+## Fixes
+ - Nether Spire structures now spawn correctly (caused by incorrect file path).
+ - Zombified Piglins now hold Golden Swords.
+
 # JDK Incompatibility Hotfix (`0.1.1+1.21`)
 ## Fixes
  - Disabled JDK preview features which caused issues with production JVMs.